# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Iron Temple API
  version: 0.1.0
  description: >
    REST API for the Iron Temple weight-lifting tracker. All weights are in
    pounds (lb).

    Authentication is a session cookie (`it_session`), issued by
    `POST /auth/login` and revoked by `POST /auth/logout`. Every endpoint
    requires it except `/health`, the `/auth/*` endpoints, and avatar images.
    Registration is open only while no account exists: the first sign-up claims
    the install and closes the door behind it.

    Programs, days and exercises are shared — the prescription is the same for
    everyone. Sessions are per-user: what you lifted is yours, and progression
    is computed from your own history alone.

    Programs are made of days (e.g. Workout A / B); each day prescribes exercises
    with sets, reps, and a starting weight. A session is a performed instance of
    a day, with one logged set per prescribed set.

    Progression is linear and computed server-side from session history:
    +5 lb per session on squat, bench, row, and overhead press; +10 lb on
    deadlift; a 10% deload after three consecutive failed sessions on a lift.
servers:
  - url: /api/v1

# Applies to every operation; the handful of public ones override with
# `security: []`.
security:
  - cookieAuth: []
...
paths:
...
  /programs:
    get:
      tags: [programs]
      operationId: listPrograms
      summary: List available programs
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgramSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
    get:
      tags: [programs]
      operationId: getProgram
      summary: Get a program with its days and prescribed exercises
      responses:
        "200":
          description: Program detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}/days/{dayId}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/DayId"
    patch:
      tags: [programs]
      operationId: updateProgramDayWeekday
      summary: Set or clear the weekday a program day is scheduled on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramDayRequest"
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}/days/{dayId}/assistance:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/DayId"
    post:
      tags: [programs]
      operationId: addAssistance
      summary: Attach an assistance exercise to a program day
      description: >
        Adds to the caller's own overlay on a shared program day; the program's
        prescription is not modified, and nobody else's view of the day changes.
        Takes effect on the next session started from this day — a session
        already in progress keeps the sets it was created with.


        Two 409s: the exercise is already on this day as assistance, or the
        program itself already prescribes it. The second is not a nicety —
        prescribing one lift twice collides on the session's own unique
        constraint, so the workout could not be started at all.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssistanceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDayAssistance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /programs/{programId}/days/{dayId}/next-session:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/DayId"
    get:
      tags: [programs]
      operationId: previewNextSession
      summary: Preview the next session for a program day
      description: >
        Returns the prescription for the day with weights computed from history
        via the program's progression rules. Does not create a session.
      responses:
        "200":
          description: Prescribed next session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescribedSession"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /sessions:
    get:
      tags: [sessions]
      operationId: listSessions
      summary: List performed sessions (most recent first)
      parameters:
        - name: programId
          in: query
          required: false
          description: Filter to sessions of a given program
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated session history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [sessions]
      operationId: createSession
      summary: Start a session from a program day
      description: >
        Materializes the day's prescription (with progression-computed weights)
        into a new session and its logged sets.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: it_session
      description: >
        Opaque session token. Set by the server as an HttpOnly, SameSite=Lax
        cookie — it is never readable by, or set from, JavaScript, so clients
        neither send nor manage it explicitly.
  parameters:
...
    ProgramId:
      name: programId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    DayId:
      name: dayId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not signed in, or the session has expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request is valid but collides with existing state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          example: session 42 not found
...
    ProgressionKind:
      type: string
      enum: [linear]
    ProgramSummary:
      type: object
      required: [id, name, description, progressionKind]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: StrongLifts 5x5
        description:
          type: string
        progressionKind:
          $ref: "#/components/schemas/ProgressionKind"
    Program:
      allOf:
        - $ref: "#/components/schemas/ProgramSummary"
        - type: object
          required: [days]
          properties:
            days:
              type: array
              items:
                $ref: "#/components/schemas/ProgramDay"
    ProgramDay:
      type: object
      required: [id, name, position, weekday, exercises, assistance]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Workout A
        position:
          type: integer
          example: 1
        weekday:
          type: integer
          nullable: true
          minimum: 0
          maximum: 6
          description: 0 = Sunday … 6 = Saturday; null if unscheduled.
        exercises:
          type: array
          description: >-
            The program's own prescription, shared by every lifter and never
            edited.
          items:
            $ref: "#/components/schemas/ProgramDayExercise"
        assistance:
          type: array
          description: >-
            The calling lifter's assistance work for this day — an additive
            overlay on the prescription above, private to them, and empty until
            they add something.
          items:
            $ref: "#/components/schemas/ProgramDayAssistance"
    ProgramDayAssistance:
      description: >-
        One assistance exercise a lifter has attached to a program day. Unlike
        ProgramDayExercise it has no starting weight and no progression: the
        prescribed weight is whatever they last logged for the lift, falling back
        to weightLb here.
      type: object
      required: [id, exerciseId, exerciseName, position, sets, reps, weightLb]
      properties:
        id:
          type: integer
          format: int64
        exerciseId:
          type: integer
          format: int64
        exerciseName:
          type: string
          example: Dip
        position:
          type: integer
          example: 1
        sets:
          type: integer
          example: 3
        reps:
          type: integer
          example: 8
        weightLb:
          type: number
          minimum: 0
          description: Fallback weight, used until the lift has been logged once.
          example: 0
    CreateAssistanceRequest:
      type: object
      required: [exerciseId, sets, reps]
      properties:
        exerciseId:
          type: integer
          format: int64
        sets:
          type: integer
          minimum: 1
          maximum: 20
          example: 3
        reps:
          type: integer
          minimum: 1
          maximum: 100
          example: 8
        weightLb:
          type: number
          minimum: 0
          default: 0
          description: Omit or send 0 for bodyweight work.
...
    UpdateProgramDayRequest:
      type: object
      required: [weekday]
      properties:
        weekday:
          type: integer
          nullable: true
          minimum: 0
          maximum: 6
          description: 0 = Sunday … 6 = Saturday; null to unschedule.
    ProgramDayExercise:
      type: object
      required:
        [id, exerciseId, exerciseName, position, sets, reps, startingWeightLb, restSeconds]
      properties:
        id:
          type: integer
          format: int64
        exerciseId:
          type: integer
          format: int64
        exerciseName:
          type: string
          example: Squat
        position:
          type: integer
          example: 1
        sets:
          type: integer
          example: 5
        reps:
          type: integer
          example: 5
        startingWeightLb:
          type: number
          minimum: 0
          example: 45
        restSeconds:
          type: integer
          description: Prescribed rest between sets (default 180 = 3 minutes)
          example: 180
    PrescribedSession:
      description: A day's prescription with progression-computed weights, not yet saved.
      type: object
      required: [programId, programDayId, programDayName, exercises]
      properties:
        programId:
          type: integer
          format: int64
        programDayId:
          type: integer
          format: int64
        programDayName:
          type: string
          example: Workout A
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/PrescribedExercise"
    ExerciseKind:
      type: string
      enum: [main, assistance]
      description: >-
        main is the program's own prescription, driven by the progression engine.
        assistance is the lifter's own addition to the day, whose weight carries
        forward from what they last logged.
    PrescribedExercise:
      type: object
      required: [exerciseId, exerciseName, kind, sets, reps, weightLb, restSeconds, progression]
      properties:
        exerciseId:
          type: integer
          format: int64
        exerciseName:
          type: string
        kind:
          $ref: "#/components/schemas/ExerciseKind"
        sets:
          type: integer
          example: 5
        reps:
          type: integer
          example: 5
        weightLb:
          type: number
          minimum: 0
          description: Computed target weight for this session
          example: 50
        restSeconds:
          type: integer
          example: 180
        progression:
          $ref: "#/components/schemas/ProgressionInfo"
    ProgressionInfo:
      description: >-
        Why the engine chose this lift's weight, so the UI can surface an
        impending stall or a deload instead of a bare number.
      type: object
      required: [status, failureCount, failuresBeforeDeload, previousWeightLb]
      properties:
        status:
          type: string
          enum: [start, advance, hold, deload, fixed]
          description: >-
            start (no history), advance (last session passed), hold (recent
            failure, repeating the weight), deload (stall hit the threshold, weight
            dropped), fixed (assistance work — no engine runs on it; the weight is
            carried forward from what was last logged).
          example: hold
        failureCount:
          type: integer
          minimum: 0
          description: Consecutive failed sessions at the current working weight.
          example: 2
        failuresBeforeDeload:
          type: integer
          minimum: 1
          description: Failures at a weight that trigger a deload.
          example: 3
        previousWeightLb:
          type: number
          minimum: 0
          description: >-
            The weight just worked — advanced past, repeated, or deloaded from;
            0 when there is no history.
          example: 185
    CreateSessionRequest:
      type: object
      required: [programDayId]
      properties:
        programDayId:
          type: integer
          format: int64
        performedOn:
          type: string
          format: date
          description: Defaults to the server's current date if omitted
...
    Session:
      type: object
      required:
        [id, programId, programName, programDayId, programDayName, performedOn, notes, createdAt, isOver, sets]
      properties:
        id:
          type: integer
          format: int64
        programId:
          type: integer
          format: int64
        programName:
          type: string
        programDayId:
          type: integer
          format: int64
        programDayName:
          type: string
          example: Workout A
        performedOn:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
          description: >
            When the lifter pressed Finish. Null if they never did — the session
            may still be over, having aged past the 12-hour cutoff.
        isOver:
          type: boolean
          description: >
            True once the session was finished by hand or was started more than
            12 hours ago. An over session is read-only and is what the
            progression engine counts.
        sets:
          type: array
          items:
            $ref: "#/components/schemas/SessionSet"
    SessionSet:
      type: object
      required:
        [id, exerciseId, exerciseName, kind, setNumber, targetReps, weightLb, completed, restSeconds]
      properties:
        id:
          type: integer
          format: int64
        exerciseId:
          type: integer
          format: int64
        exerciseName:
          type: string
          example: Squat
        kind:
          $ref: "#/components/schemas/ExerciseKind"
        setNumber:
          type: integer
          example: 1
        targetReps:
          type: integer
          example: 5
        actualReps:
          type: integer
          nullable: true
          minimum: 0
          description: Null until the set is logged
        weightLb:
          type: number
          minimum: 0
          example: 50
        completed:
          type: boolean
        restSeconds:
          type: integer
          example: 180
    SessionExerciseWeight:
      description: An exercise's set/rep prescription and top weight for a session.
      type: object
      required: [exerciseName, sets, reps, weightLb]
      properties:
        exerciseName:
          type: string
        sets:
          type: integer
        reps:
          type: integer
        weightLb:
          type: number
          minimum: 0
    SessionSummary:
      type: object
      required:
        [id, programId, programName, programDayId, programDayName, performedOn, setCount, completedSetCount, volumeLb, isOver, exercises]
      properties:
        id:
          type: integer
          format: int64
        programId:
          type: integer
          format: int64
        programName:
          type: string
        programDayId:
          type: integer
          format: int64
        programDayName:
          type: string
        performedOn:
          type: string
          format: date
        setCount:
          type: integer
        completedSetCount:
          type: integer
        volumeLb:
          type: number
          minimum: 0
          example: 8450
          description: >
            Total weight moved in this session — the sum of actualReps × weightLb
            over every set with reps logged. A set logged short of its target
            still counts what was lifted, so this does not track completedSetCount.
        isOver:
          type: boolean
          description: >
            True once the session was finished by hand or was started more than
            12 hours ago.
        exercises:
          type: array
          description: Per-exercise top weights, in prescription order.
          items:
            $ref: "#/components/schemas/SessionExerciseWeight"
    SessionList:
      type: object
      required: [items, total, totalVolumeLb, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
        total:
          type: integer
        totalVolumeLb:
          type: number
          minimum: 0
          example: 412650
          description: >
            Volume summed over every session matching the filter, not just the
            ones on this page — a lifetime total that does not change as the
            caller pages through the history.
        limit:
          type: integer
        offset:
          type: integer

    # ---- Racked ----
    # The recap. Every statistic is computed server-side so that the page and
    # the monthly email render the same numbers from the same report; nothing
    # here is aggregated again on the client.
...